+++ sdk/datafactory/azure-resourcemanager-datafactory/src/main/java/com/azure/resourcemanager/datafactory/models/TabularSource.java
@@ -78,7 +78,9 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
     @JsonSubTypes.Type(name = "DynamicsAXSource", value = DynamicsAXSource.class),
     @JsonSubTypes.Type(name = "OracleServiceCloudSource", value = OracleServiceCloudSource.class),
     @JsonSubTypes.Type(name = "GoogleAdWordsSource", value = GoogleAdWordsSource.class),
-    @JsonSubTypes.Type(name = "AmazonRedshiftSource", value = AmazonRedshiftSource.class) })
+    @JsonSubTypes.Type(name = "AmazonRedshiftSource", value = AmazonRedshiftSource.class),
+    @JsonSubTypes.Type(name = "WarehouseSource", value = WarehouseSource.class),
+    @JsonSubTypes.Type(name = "SalesforceV2Source", value = SalesforceV2Source.class) })
 @Fluent
 public class TabularSource extends CopySource {
     /*