+++ sdk/streamanalytics/azure-resourcemanager-streamanalytics/src/main/java/com/azure/resourcemanager/streamanalytics/models/FunctionsCreateOrReplaceHeaders.java
@@ -5,24 +5,34 @@
 package com.azure.resourcemanager.streamanalytics.models;
 
 import com.azure.core.annotation.Fluent;
-import com.azure.core.util.logging.ClientLogger;
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.azure.core.http.HttpHeaderName;
+import com.azure.core.http.HttpHeaders;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-/** The FunctionsCreateOrReplaceHeaders model. */
+/**
+ * The FunctionsCreateOrReplaceHeaders model.
+ */
 @Fluent
 public final class FunctionsCreateOrReplaceHeaders {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(FunctionsCreateOrReplaceHeaders.class);
-
     /*
      * The Etag property.
      */
     @JsonProperty(value = "Etag")
     private String etag;
 
+    // HttpHeaders containing the raw property values.
+    /**
+     * Creates an instance of FunctionsCreateOrReplaceHeaders class.
+     * 
+     * @param rawHeaders The raw HttpHeaders that will be used to create the property values.
+     */
+    public FunctionsCreateOrReplaceHeaders(HttpHeaders rawHeaders) {
+        this.etag = rawHeaders.getValue(HttpHeaderName.ETAG);
+    }
+
     /**
      * Get the etag property: The Etag property.
-     *
+     * 
      * @return the etag value.
      */
     public String etag() {
@@ -31,7 +41,7 @@ public final class FunctionsCreateOrReplaceHeaders {
 
     /**
      * Set the etag property: The Etag property.
-     *
+     * 
      * @param etag the etag value to set.
      * @return the FunctionsCreateOrReplaceHeaders object itself.
      */
@@ -42,7 +52,7 @@ public final class FunctionsCreateOrReplaceHeaders {
 
     /**
      * Validates the instance.
-     *
+     * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {