+++ sdk/vision/azure-ai-vision-imageanalysis/src/main/java/com/azure/ai/vision/imageanalysis/ImageAnalysisClient.java
@@ -5,6 +5,7 @@ package com.azure.ai.vision.imageanalysis;
 
 import com.azure.ai.vision.imageanalysis.implementation.ImageAnalysisClientImpl;
 import com.azure.ai.vision.imageanalysis.implementation.models.ImageUrl;
+import com.azure.ai.vision.imageanalysis.models.ImageAnalysisOptions;
 import com.azure.ai.vision.imageanalysis.models.ImageAnalysisResult;
 import com.azure.ai.vision.imageanalysis.models.VisualFeatures;
 import com.azure.core.annotation.Generated;
@@ -27,6 +28,18 @@ import java.util.stream.Collectors;
 
 /**
  * Initializes a new instance of the synchronous ImageAnalysisClient type.
+ *
+ * <!-- src_embed com.azure.ai.vision.imageanalysis.sync-client -->
+ * <pre>
+ * &#47;&#47;
+ * &#47;&#47; Create a synchronous Image Analysis client.
+ * &#47;&#47;
+ * ImageAnalysisClient client = new ImageAnalysisClientBuilder&#40;&#41;
+ *     .endpoint&#40;endpoint&#41;
+ *     .credential&#40;new KeyCredential&#40;key&#41;&#41;
+ *     .buildClient&#40;&#41;;
+ * </pre>
+ * <!-- end com.azure.ai.vision.imageanalysis.sync-client -->
  */
 @ServiceClient(builder = ImageAnalysisClientBuilder.class)
 public final class ImageAnalysisClient {
@@ -36,7 +49,7 @@ public final class ImageAnalysisClient {
 
     /**
      * Initializes an instance of ImageAnalysisClient class.
-     * 
+     *
      * @param serviceClient the service client implementation.
      */
     @Generated
@@ -63,8 +76,7 @@ public final class ImageAnalysisClient {
      * <td>No</td>
      * <td>The desired language for result generation (a two-letter language code).
      * If this option is not specified, the default value 'en' is used (English).
-     * See https://aka.ms/cv-languages for a list of supported languages.
-     * At the moment, only tags can be generated in none-English languages.</td>
+     * See https://aka.ms/cv-languages for a list of supported languages.</td>
      * </tr>
      * <tr>
      * <td>gender-neutral-caption</td>
@@ -194,7 +206,7 @@ public final class ImageAnalysisClient {
      *     }
      * }
      * }</pre>
-     * 
+     *
      * @param visualFeatures A list of visual features to analyze.
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
@@ -232,8 +244,7 @@ public final class ImageAnalysisClient {
      * <td>No</td>
      * <td>The desired language for result generation (a two-letter language code).
      * If this option is not specified, the default value 'en' is used (English).
-     * See https://aka.ms/cv-languages for a list of supported languages.
-     * At the moment, only tags can be generated in none-English languages.</td>
+     * See https://aka.ms/cv-languages for a list of supported languages.</td>
      * </tr>
      * <tr>
      * <td>gender-neutral-caption</td>
@@ -365,7 +376,7 @@ public final class ImageAnalysisClient {
      *     }
      * }
      * }</pre>
-     * 
+     *
      * @param visualFeatures A list of visual features to analyze.
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
@@ -386,7 +397,7 @@ public final class ImageAnalysisClient {
 
     /**
      * Performs a single Image Analysis operation.
-     * 
+     *
      * @param visualFeatures A list of visual features to analyze.
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
@@ -394,7 +405,6 @@ public final class ImageAnalysisClient {
      * @param language The desired language for result generation (a two-letter language code).
      * If this option is not specified, the default value 'en' is used (English).
      * See https://aka.ms/cv-languages for a list of supported languages.
-     * At the moment, only tags can be generated in none-English languages.
      * @param genderNeutralCaption Boolean flag for enabling gender-neutral captioning for Caption and Dense Captions
      * features.
      * By default captions may contain gender terms (for example: 'man', 'woman', or 'boy', 'girl').
@@ -444,7 +454,7 @@ public final class ImageAnalysisClient {
 
     /**
      * Performs a single Image Analysis operation.
-     * 
+     *
      * @param visualFeatures A list of visual features to analyze.
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
@@ -469,7 +479,7 @@ public final class ImageAnalysisClient {
 
     /**
      * Performs a single Image Analysis operation.
-     * 
+     *
      * @param visualFeatures A list of visual features to analyze.
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
@@ -477,7 +487,6 @@ public final class ImageAnalysisClient {
      * @param language The desired language for result generation (a two-letter language code).
      * If this option is not specified, the default value 'en' is used (English).
      * See https://aka.ms/cv-languages for a list of supported languages.
-     * At the moment, only tags can be generated in none-English languages.
      * @param genderNeutralCaption Boolean flag for enabling gender-neutral captioning for Caption and Dense Captions
      * features.
      * By default captions may contain gender terms (for example: 'man', 'woman', or 'boy', 'girl').
@@ -527,7 +536,7 @@ public final class ImageAnalysisClient {
 
     /**
      * Performs a single Image Analysis operation.
-     * 
+     *
      * @param visualFeatures A list of visual features to analyze.
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
@@ -558,51 +567,107 @@ public final class ImageAnalysisClient {
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
      * @param options Additional image analysis options.
-     * 
+     *
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws HttpResponseException thrown if the request is rejected by server.
      * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
      * @throws ResourceNotFoundException thrown if the request is rejected by server on status code 404.
      * @throws ResourceModifiedException thrown if the request is rejected by server on status code 409.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * 
+     *
      * @return represents the outcome of an Image Analysis operation.
+     * 
+     * <!-- src_embed com.azure.ai.vision.imageanalysis.sync-analysis-from-url -->
+     * <pre>
+     * &#47;&#47;
+     * &#47;&#47; Synchronous analysis of an image file &#40;https:&#47;&#47;aka.ms&#47;azsdk&#47;image-analysis&#47;sample.jpg&#41;,
+     * &#47;&#47; using all visual features, with all options set. You will likely need fewer visual features
+     * &#47;&#47; and only some &#40;or no&#41; options set.
+     * &#47;&#47;
+     * ImageAnalysisOptions options = new ImageAnalysisOptions&#40;&#41;
+     *     .setLanguage&#40;&quot;en&quot;&#41;
+     *     .setGenderNeutralCaption&#40;true&#41;
+     *     .setSmartCropsAspectRatios&#40;Arrays.asList&#40;0.9, 1.33&#41;&#41;
+     *     .setModelVersion&#40;&quot;latest&quot;&#41;;
+     *
+     * ImageAnalysisResult result = client.analyze&#40;
+     *     new URL&#40;&quot;https:&#47;&#47;aka.ms&#47;azsdk&#47;image-analysis&#47;sample.jpg&quot;&#41;,
+     *     Arrays.asList&#40;
+     *         VisualFeatures.SMART_CROPS,
+     *         VisualFeatures.CAPTION,
+     *         VisualFeatures.DENSE_CAPTIONS,
+     *         VisualFeatures.OBJECTS,
+     *         VisualFeatures.PEOPLE,
+     *         VisualFeatures.READ,
+     *         VisualFeatures.TAGS&#41;,
+     *     options&#41;;
+     * </pre>
+     * <!-- end com.azure.ai.vision.imageanalysis.sync-analysis-from-url -->
      */
+    @ServiceMethod(returns = ReturnType.SINGLE)
     public ImageAnalysisResult analyze(URL imageUrl, List<VisualFeatures> visualFeatures,
         ImageAnalysisOptions options) {
         if (options == null) {
             return analyzeFromUrl(visualFeatures, new ImageUrl(imageUrl.toString()), null, null, null, null);
         } else {
             return analyzeFromUrl(visualFeatures, new ImageUrl(imageUrl.toString()), options.getLanguage(),
-                options.getGenderNeutralCaption(), options.getSmartCropsAspectRatios(), options.getModelVersion());
+                options.isGenderNeutralCaption(), options.getSmartCropsAspectRatios(), options.getModelVersion());
         }
     }
 
     /**
      * Performs a single Image Analysis operation on a give image buffer.
      *
-     * @param imageBuffer The image to analyze.
+     * @param imageData The image to analyze.
      * @param visualFeatures A list of visual features to analyze.
      * Seven visual features are supported: Caption, DenseCaptions, Read (OCR), Tags, Objects, SmartCrops, and People.
      * At least one visual feature must be specified.
      * @param options Additional image analysis options.
-     * 
+     *
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws HttpResponseException thrown if the request is rejected by server.
      * @throws ClientAuthenticationException thrown if the request is rejected by server on status code 401.
      * @throws ResourceNotFoundException thrown if the request is rejected by server on status code 404.
      * @throws ResourceModifiedException thrown if the request is rejected by server on status code 409.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * 
+     *
      * @return represents the outcome of an Image Analysis operation.
+     * 
+     * <!-- src_embed com.azure.ai.vision.imageanalysis.sync-analysis-from-file -->
+     * <pre>
+     * &#47;&#47;
+     * &#47;&#47; Synchronous analysis of an image file &#40;sample.jpg&#41;, using all visual features,
+     * &#47;&#47; with all options set. You will likely need fewer visual features and only
+     * &#47;&#47; some &#40;or no&#41; options set.
+     * &#47;&#47;
+     * ImageAnalysisOptions options = new ImageAnalysisOptions&#40;&#41;
+     *     .setLanguage&#40;&quot;en&quot;&#41;
+     *     .setGenderNeutralCaption&#40;true&#41;
+     *     .setSmartCropsAspectRatios&#40;Arrays.asList&#40;0.9, 1.33&#41;&#41;
+     *     .setModelVersion&#40;&quot;latest&quot;&#41;;
+     *
+     * ImageAnalysisResult result = client.analyze&#40;
+     *     BinaryData.fromFile&#40;new File&#40;&quot;sample.jpg&quot;&#41;.toPath&#40;&#41;&#41;,
+     *     Arrays.asList&#40;
+     *         VisualFeatures.SMART_CROPS,
+     *         VisualFeatures.CAPTION,
+     *         VisualFeatures.DENSE_CAPTIONS,
+     *         VisualFeatures.OBJECTS,
+     *         VisualFeatures.PEOPLE,
+     *         VisualFeatures.READ,
+     *         VisualFeatures.TAGS&#41;,
+     *     options&#41;;
+     * </pre>
+     * <!-- end com.azure.ai.vision.imageanalysis.sync-analysis-from-file -->
      */
-    public ImageAnalysisResult analyze(BinaryData imageBuffer, List<VisualFeatures> visualFeatures,
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public ImageAnalysisResult analyze(BinaryData imageData, List<VisualFeatures> visualFeatures,
         ImageAnalysisOptions options) {
         if (options == null) {
-            return analyzeFromBuffer(visualFeatures, imageBuffer, null, null, null, null);
+            return analyzeFromBuffer(visualFeatures, imageData, null, null, null, null);
         } else {
-            return analyzeFromBuffer(visualFeatures, imageBuffer, options.getLanguage(),
-                options.getGenderNeutralCaption(), options.getSmartCropsAspectRatios(), options.getModelVersion());
+            return analyzeFromBuffer(visualFeatures, imageData, options.getLanguage(), options.isGenderNeutralCaption(),
+                options.getSmartCropsAspectRatios(), options.getModelVersion());
         }
     }
 }